# Improve performance of task card