--- a/app/styles/components/task-card.scss
+++ b/app/styles/components/task-card.scss
@@ -20,6 +20,7 @@
   }
 
   &__assignee {
+    display: flex;
     flex-direction: row;
     justify-content: flex-start;
     margin: 0;