# What's in this PR?

This is a spike and should not (yet) be considered production-ready code.

The changes made here have so far improved render of a given task list from ~1500ms to ~500ms, resulting in a 3x performance improvement.

- Replaced `svg/sprite-icon` component on `task-card` with simple HTML
  - removes ~1ms render multiplied by `n` cards
- Does not render `ember-power-select` in `task-assignment` until hovering the card
  - this may have implications for mobile usage that needs considered
  - removes ~10ms render multiplied by `n` cards
- Does not render the `github/issue-link` component until hovering the card and instead renders simple HTML
  - removes < 1ms render multiplied by `n` cards

The `tooltip-on-component` from `ember-tooltip` has similarly deleterious effects on performance when rendering across multiple cards.

One thing that we could do is to change the behavior from hovering to whether the card is in the viewport (and could possibly unrender the complex elements when leaving the viewport, though this unrendering does not work for `mouseLeave` because of other UX considerations).

Much of the mobile behavior changes may depend on whether we can reliably sniff the mobile device. Ideally the behavior would be the same across devices and screen sizes.

Fixes #1504