@JoshSmith 2017-11-07
@begedin I could use some thoughts on this.