@JoshSmith 2017-11-07
It may be worth also testing removing our use of `ember-composable-helpers` and the like for simple computed properties.