@begedin 2017-11-08
> Does not render ember-power-select in task-assignment until hovering the card
> - this may have implications for mobile usage that needs considered

Viewport might work, with probably a moderate amount of complexity. Another option might be to somehow trigger rendering followed by the first open on click. Something like 

```Javascript
click() {
  set(this, 'renderDropdown', true);
  run(this, 'afterRender', () => get(this, 'showDropdown')()
}
```
> Does not render the github/issue-link component until hovering the card and instead renders simple HTML

Similar approach might work here, but not sure.